split allocator and archost to separate files
